requirements.txt
# Required packages for Data Strategy Recommendation Bot
openai>=1.0.0
qdrant-client>=1.7.0
click>=8.1.0
pydantic>=2.0.0
rich>=13.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

